apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.8",
    "psycopg2==2.9.10",
    "pydantic==2.10.6",
    "pyjwt>=2.10.1",
    "python-dotenv==1.0.1",
    "requests>=2.32.3",
    "uvicorn==0.34.0",
]

[dependency-groups]
dev = [
    "mypy==1.15.0",
    "pre-commit==4.1.0",
    "ruff==0.9.5",
]
